ajouter un site

+++ controllers/SiteController.php
@@ -9,50 +9,52 @@ use App\Models\Reservation;
 
 use App\Providers\View;
 use App\Providers\Validator;
+use App\Providers\Auth;
 
 class SiteController
 {
     public function index()
     {
         $site = new Site;
-        $select = $site->select();
+        $select = $site->select('id', 'desc');
 
         return View::render("site/index", ['sites' => $select, 'ASSET' => ASSET, 'base' => BASE]);
     }
 
-    public function create($data = [])
-    { icii
+    public function create()
+    {
+        Auth::session();
+        Auth::privilege(1);
+
         $categorie = new Categorie;
         $selectCategorie = $categorie->select();
-        $collaborateur = new Collaborateur;
-        $selectId = $collaborateur->selectId($data['id']);
-        if ($selectId) {
-            return View::render("site/create", ['site' => $selectId]);
-        } else {
-            return View::render('error', ['msg' => 'Collaborateur(id) non trouvé dans la Base de données!']);
-        }
+
+        return View::render("site/create", ['categories' => $selectCategorie]);
     }
 
     public function store($data)
     {
+        Auth::session();
         $validator = new Validator;
         $validator->field('siteNom', $data['siteNom'])->required();
         $validator->field('siteDescription', $data['siteDescription'])->required();
-        $validator->field('capacite', $data['capacite'])->int()->max(10);
+        $validator->field('capacite', $data['capacite'])->int()->bigger(10);
         $validator->field('prix', $data['prix'])->int()->max(10);
         $validator->field('urlImage', $data['urlImage'])->required();
-        $validator->field('categorieId', $data['categorieId'])->int()->max(1);
+        $validator->field('categorieId', $data['categorieId'])->required();
 
         if ($validator->isSuccess()) {
             $Site = new Site;
             $insert = $Site->insert($data);
-            return View::redirect('site/show?id=' . $data['id']);
+            return View::redirect('sites');
         } else {
             $errors = $validator->getErrors();
             $Site = new Site;
             $categorie = new Categorie;
             $selectId = $categorie->selectId($data['categorieId']);
-            return View::render('site/create', ['errors' => $errors, 'categorie' => $selectId, 'Site' => $data]);
+            $selectCategorie = $categorie->select();
+
+            return View::render('site/create', ['errors' => $errors, 'categorie' => $selectId, 'categories' => $selectCategorie, 'site' => $data]);
         }
     }
 
@@ -78,6 +80,7 @@ class SiteController
 
     public function edit($data = [])
     {
+        Auth::session();
         if (isset($data['id']) && $data['id'] != null) {
             $Site = new Site;
             $selectId = $Site->selectId($data['id']);
@@ -97,10 +100,11 @@ class SiteController
 
     public function update($data = [], $get = [])
     {
+        Auth::session();
         if (isset($get['id']) && $get['id'] != null) {
             $validator = new Validator;
             $validator->field('siteNom', $data['siteNom'])->max(20);
-            $validator->field('siteDescription', $data['siteDescription'])->max(50);
+            $validator->field('siteDescription', $data['siteDescription'])->max(100);
             $validator->field('capacite', $data['capacite'])->int()->bigger(10);
             $validator->field('prix', $data['prix'])->int();
             $validator->field('urlImage', $data['urlImage']);
@@ -133,24 +137,26 @@ class SiteController
 
     public function delete($data)
     {
-        $reservation = new Reservation;
-        $selectListe = $reservation->selectListe('siteId', $data['id']);
-        $count = count($selectListe);
+        if (Auth::session()) {
+            $reservation = new Reservation;
+            $selectListe = $reservation->selectListe('siteId', $data['id']);
+            $count = count($selectListe);
 
-        if ($count > 0) {
-            foreach ($selectListe as $selected) {
-                $reservation = new Reservation;
-                $delete = $reservation->delete($selected['id']);
+            if ($count > 0) {
+                foreach ($selectListe as $selected) {
+                    $reservation = new Reservation;
+                    $delete = $reservation->delete($selected['id']);
+                }
             }
-        }
 
-        $Site = new Site;
-        $delete = $Site->delete($data['id']);
+            $Site = new Site;
+            $delete = $Site->delete($data['id']);
 
-        if ($delete) {
-            View::redirect('sites');
-        } else {
-            return View::render('error', ['msg' => 'Echec de la suppression!']);
+            if ($delete) {
+                View::redirect('sites');
+            } else {
+                return View::render('error', ['msg' => 'Echec de la suppression!']);
+            }
         }
     }
 }
